test(web): split dataset service server tests
